refactor: 관심사에 따른 Group Controller 분리
